clean(vod): Cleans vod file structure
