+++ Sources/Live/ConcreteImplementation/VOD/JSONVodItemType.swift
@@ -1,0 +1,13 @@
+//
+//  JSONVodItemType.swift
+//  
+//
+//  Created by Sacha DSO on 27/11/2021.
+//
+
+import Foundation
+
+public enum JSONVodItemType: String, Codable {
+	case video
+	case playlist
+}
